Changed folder structure and put the states in their respective namespace to keep states only accassable for those who require them
