+++ Assets/_Core/Scripts/Specific/Game/GameView.cs
@@ -1,4 +1,5 @@
 ﻿using UnityEngine;
+using SurvivalGame;
 
 public class GameView : MonoBaseView
 {
