+++ Assets/_Core/Scripts/Specific/Game/States/SurvivalGameState/SurvivalGameState.cs
@@ -1,38 +1,41 @@
-﻿public class SurvivalGameState : BaseGameState, IGame
+﻿namespace SurvivalGame
 {
-	public TurretModel TurretModel
+	public class SurvivalGameState : BaseGameState, IGame
 	{
-		get; private set;
-	}
+		public TurretModel TurretModel
+		{
+			get; private set;
+		}
 
-	public WordsDisplayerModel WordsDisplayerModel
-	{
-		get; private set;
-	}
+		public WordsDisplayerModel WordsDisplayerModel
+		{
+			get; private set;
+		}
 
-	public GameStateManager<SurvivalGameState> SurvivalGameStateManager
-	{
-		get; private set;
-	}
+		public GameStateManager<SurvivalGameState> SurvivalGameStateManager
+		{
+			get; private set;
+		}
 
-	protected override void OnSetupState()
-	{
-		SurvivalGameStateManager = new GameStateManager<SurvivalGameState>(this);
+		protected override void OnSetupState()
+		{
+			SurvivalGameStateManager = new GameStateManager<SurvivalGameState>(this);
 
-		// Setup UI
-		WordsDisplayerModel = new WordsDisplayerModel(Game.TimekeeperModel);
+			// Setup UI
+			WordsDisplayerModel = new WordsDisplayerModel(Game.TimekeeperModel);
 
-		// Setup Player
-		TurretModel = new TurretModel(Game.TimekeeperModel);
-	}
+			// Setup Player
+			TurretModel = new TurretModel(Game.TimekeeperModel);
+		}
 
-	protected override void OnStartState()
-	{
-		SurvivalGameStateManager.SetGameState<IntroGameState>();
-	}
+		protected override void OnStartState()
+		{
+			SurvivalGameStateManager.SetGameState<IntroGameState>();
+		}
 
-	protected override void OnEndState()
-	{
+		protected override void OnEndState()
+		{
 
+		}
 	}
 }
