--- Assets/_Core/Scripts/Specific/Game/States/WavesGameState/WavesGameStateView.cs
@@ -1,26 +1,0 @@
-﻿using UnityEngine;
-
-public class WavesGameStateView : MonoGameStateView<SurvivalGameState>
-{
-	private WavesGameState _wavesGameState;
-
-	[SerializeField]
-	private WaveSystemView _waveSystemView;
-
-	protected override void OnPreStartStateView()
-	{
-		_wavesGameState = GameState as WavesGameState;
-
-		Controller.Link(_wavesGameState.WaveSystem, _waveSystemView);
-	}
-
-	protected override void OnStartStateView()
-	{
-
-	}
-
-	protected override void OnEndStateView()
-	{
-		_wavesGameState = null;
-	}
-}
